+++ app/src/ui/screens/scr_welcome.c
@@ -6,7 +6,6 @@
 #include <zephyr/smf.h>
 #include <app_version.h>
 
-#include "sys_sm_module.h"
 #include "hw_module.h"
 
 #include "display_module.h"
@@ -21,7 +20,7 @@ extern lv_style_t style_h2;
 extern lv_style_t style_info;
 extern lv_style_t style_icon;
 
-extern int curr_screen;
+extern uint8_t curr_screen;
 
 static void anim_x_cb(void *var, int32_t v)
 {
